lesson05-phancaominhnhat 10.06.2021

+++ lesson05-phancaominhnhat/src/bai5/SachThaoKhao.java
@@ -1,18 +1,18 @@
 package bai5;
 
 public class SachThaoKhao extends Sach {
-	public SachThaoKhao(String maSach, int donGia, String nhaXuatBan, int thue) {
+	public SachThaoKhao(String maSach, int donGia, String nhaXuatBan, Double thue) {
 		super(maSach, donGia, nhaXuatBan);
 		this.thue = thue;
 	}
 
-	private int thue;
+	private Double thue;
 
-	public int getThue() {
+	public Double getThue() {
 		return thue;
 	}
 
-	public void setThue(int thue) {
+	public void setThue(Double thue) {
 		this.thue = thue;
 	}
